% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatBank-GetData.R
\name{StatBank.GetData}
\alias{StatBank.GetData}
\title{Get Data from Statistics Denmarks}
\usage{
StatBank.GetData(TableId, lang = "DA", variables = NULL)
}
\arguments{
\item{TableId}{character("TableId"); Table id to extract metadata from - get a list of TableId using \code{\link{StatBank.GetTabels}}}

\item{lang}{character("DA", "ENG); Language to export Danish "DA" (default) or English ENG}

\item{variables}{list(Variableid = c(Value-ids)); list of variable or variables to include in the API call - see details}
}
\value{
Returns a Data Frame
}
\description{
Retrieves metadata about a dataset. The result contains eg. information about variables and their values used when extracting data.
}
\details{
To include variables to the API call build a list, where each object in the list is named with the Variableid and value added as character or character list. Use \code{\link{StatBank.GetMetadata}} to get a list of variables or browse http://www.statistikbanken.dk/.
}
\examples{
#Building a list to get variables indto the call - note that * includes all values of that variable
variablelist <- list(UDDANNELSE = c("H40", "H4020", "H402015", "H4024"), FSTATUS = "*")

StatBank.GetData("UDDAKT40", variables = variablelist)
}

