% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatBank-GetMetadata.R
\name{StatBank.GetMetadata}
\alias{StatBank.GetMetadata}
\title{Get metadata for a tabel from Statistics Denmarks}
\usage{
StatBank.GetMetadata(TableId, lang = "DA")
}
\arguments{
\item{TableId}{character("TableId"); Table id to extract metadata from - get a list of TableId using \code{\link{StatBank.GetTabels}}}

\item{lang}{character("DA", "ENG); Language to export Danish "DA" (default) or English ENG}
}
\value{
Returns a list
}
\description{
Retrieves metadata about a dataset. The result contains eg. information about variables and their values used when extracting data.
}

