% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatBank-GetTabels.R
\name{StatBank.GetTabels}
\alias{StatBank.GetTabels}
\title{Get a list of tabels from Statistics Denmarks}
\usage{
StatBank.GetTabels(discontinued = FALSE)
}
\arguments{
\item{discontinued}{logical; FALSE (default) - Include discontinued tables}
}
\value{
Returns a Data Frame
}
\description{
Retrieves list of datasets. Filtering by subject is done by using the codes from the "subjects"-call.
}

